plugs: where a turn ends on a live line, said once
A phone line hands over no whole utterance and signals no end of turn, so the phone plug reads both out of the audio: speech while the samples carry it, the turn over once the line has been quiet long enough, and a far end that never speaks answered without words. Every plug over a live line has exactly that problem, and the next one is a room rather than a call, so the reading moves out of the phone plug into a module both can hold.

No behavior changes: same thresholds, same audio, same turns.
